requirements.txt
py-postgresql==1.2.1
